build: remove pkg, use esbuild instead
