+++ package.json
@@ -33,8 +33,7 @@
     "prettier": "yarn run prettier:base --write",
     "prettier:check": "yarn run prettier:base --check",
     "release": "lerna version",
-    "pkg": "lerna run --stream build && lerna run --stream pkg",
-    "pkg:local": "lerna run --stream build && lerna run --stream pkg:local"
+    "esbuild": "cd packages/www && yarn run static && cd ../api && yarn run esbuild"
   },
   "devDependencies": {
     "@commitlint/cli": "^16.2.3",
