+++ packages/api/package.json
@@ -36,8 +36,6 @@
     "test:dev": "jest \"${PWD}/src\" -i --silent --watch",
     "test:build": "parcel build --no-autoinstall --no-minify --bundle-node-modules -t browser --out-dir ../dist-worker ../src/worker.js",
     "coverage": "yarn run test --coverage",
-    "pkg": "pkg --out-path=bin --no-bytecode --public-packages \"*\" --public . --compress GZip ",
-    "pkg:local": "yarn run prepare && yarn run pkg --targets=$(node -p '`node18-${process.env.GOOS ?? os.platform()}-${process.env.GOARCH ?? os.arch()}`')",
     "esbuild": "yarn run compile-schemas && node esbuild.mjs"
   },
   "dependencies": {
@@ -145,7 +143,6 @@
     "npm-run-all": "^4.1.5",
     "nyc": "^14.0.0",
     "parcel": "^1.12.4",
-    "pkg": "^5.8.1",
     "prettier": "^2.6.2",
     "redoc-cli": "^0.8.3",
     "tus-js-client": "^3.0.0-0",
@@ -163,28 +160,11 @@
     ]
   },
   "browserslist": [
-    "node 10"
+    "node 16"
   ],
   "types": "src/schema/types.d.ts",
   "nodemonConfig": {
     "delay": "250"
-  },
-  "pkg": {
-    "targets": [
-      "node18-macos-arm64",
-      "node18-macos-x64",
-      "node18-linux-arm64",
-      "node18-linux-x64",
-      "node18-windows-x64"
-    ],
-    "assets": [
-      "../www/.next",
-      "../www/next.config.js",
-      "../../node_modules/dom-helpers/**/*",
-      "../../node_modules/react-icons/**/*",
-      "../../node_modules/next-mdx-remote/**/*",
-      "../../node_modules/livekit-server-sdk/**/*"
-    ]
   },
   "jest": {
     "testTimeout": 30000
